Rename package to presentationmodel
