+++ presentation-model/src/main/java/com/iluwatar/presentationmodel/DisplayedAlbums.java
@@ -22,40 +22,49 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package com.iluwatar.presentation;
+package com.iluwatar.presentationmodel;
 
-import org.junit.jupiter.api.Test;
+import java.util.ArrayList;
+import java.util.List;
 
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertTrue;
+import lombok.Getter;
+import lombok.extern.slf4j.Slf4j;
 
-class AlbumTest {
-  @Test
-  void testSetTitle(){
-    Album album = new Album("a", "b", false, "");
-    album.setTitle("b");
-    assertEquals("b", album.getTitle());
+/**
+ * a class used to deal with albums.
+ *
+ */
+@Slf4j
+@Getter
+public class DisplayedAlbums {
+  /**
+   * albums a list of albums.
+   */
+  private final List<Album> albums;
+
+  /**
+   * a constructor method.
+   */
+  public DisplayedAlbums() {
+    this.albums = new ArrayList<>();
   }
 
-  @Test
-  void testSetArtist(){
-    Album album = new Album("a", "b", false, "");
-    album.setArtist("c");
-    assertEquals("c", album.getArtist());
-  }
-
-  @Test
-  void testSetClassical(){
-    Album album = new Album("a", "b", false, "");
-    album.setClassical(true);
-    assertTrue(album.isClassical());
-  }
-
-  @Test
-  void testSetComposer(){
-    Album album = new Album("a", "b", false, "");
-    album.setClassical(true);
-    album.setComposer("w");
-    assertEquals("w", album.getComposer());
+  /**
+   * a method used to add a new album to album list.
+   *
+   * @param title       the title of the album.
+   * @param artist      the artist name of the album.
+   * @param isClassical is the album classical, true or false.
+   * @param composer    only when the album is classical,
+   *                    composer can have content.
+   */
+  public void addAlbums(final String title,
+                        final String artist, final boolean isClassical,
+                        final String composer) {
+    if (isClassical) {
+      this.albums.add(new Album(title, artist, true, composer));
+    } else {
+      this.albums.add(new Album(title, artist, false, ""));
+    }
   }
 }
