+++ presentation-model/src/main/java/com/iluwatar/presentationmodel/PresentationModel.java
@@ -22,7 +22,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package com.iluwatar.presentation;
+package com.iluwatar.presentationmodel;
 
 import lombok.extern.slf4j.Slf4j;
