+++ presentation-model/src/test/java/com/iluwatar/presentationmodel/DisplayedAlbumsTest.java
@@ -22,22 +22,25 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package com.iluwatar.presentation;
+package com.iluwatar.presentationmodel;
 
 import org.junit.jupiter.api.Test;
 
-import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
+import static org.junit.jupiter.api.Assertions.assertEquals;
 
-/**
- * Issue: Add at least one assertion to this test case.
- *
- * Solution: Inserted assertion to check whether the execution of the main method in {@link App}
- * throws an exception.
- */
-class AppTest {
+class DisplayedAlbumsTest {
+  @Test
+  void testAdd_true(){
+    DisplayedAlbums displayedAlbums = new DisplayedAlbums();
+    displayedAlbums.addAlbums("title", "artist", true, "composer");
+    assertEquals("composer", displayedAlbums.getAlbums().get(0).getComposer());
 
-    @Test
-    void shouldExecuteApplicationWithoutException() {
-        assertDoesNotThrow(() -> App.main(new String[]{}));
-    }
+  }
+
+  @Test
+  void testAdd_false(){
+    DisplayedAlbums displayedAlbums = new DisplayedAlbums();
+    displayedAlbums.addAlbums("title", "artist", false, "composer");
+    assertEquals("", displayedAlbums.getAlbums().get(0).getComposer());
+  }
 }
